adding in env vars related to the link from FC to reporting services
FARM_CALENDAR_REPORTING_API_ROOT will be set by default (example.env) to
the reporting api root url, and the FARM_CALENDAR_DISABLE_REPORTING flag
will be set to False by default. If the user wants to disable the report
option in the farmcalendar this flag can be passed as true instead. For
the Cloud deployment the FARM_CALENDAR_REPORTING_API_ROOT also needs to
be changed to the actual external DNS for the reporting service (not
localhost). but we are leaving the conf set for localhost easy setup as
the defaul in the example.env